refactor: refactor score screen and providers for improved structure and functionality
- Moved score-related utility functions and constants to score_view_helpers.dart.
- Created score_screen_actions.dart for handling score refresh and last updated caching.
- Updated ScoreScreen to utilize new helper functions and actions for better readability and maintainability.
- Enhanced academic performance provider to streamline data loading and caching logic.
